ai/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
openai-whisper==20231117
python-multipart==0.0.6
aiofiles==23.2.1
yt-dlp==2025.3.31
requests==2.32.3
